+++ frontend/index.tsx
@@ -1,12 +1,8 @@
 import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import { App } from './components/App';
+import './global.scss';
 
 window.process = { cwd: () => '' } as NodeJS.Process;
 
-ReactDOM.render(
-  <Suspense fallback={'WAITING FOR BACKEND'}>
-    <App />
-  </Suspense>,
-  document.querySelector('#root')
-);
+ReactDOM.render(<App />, document.querySelector('#root'));